+++ src/settings/MappingProfiles/ViewMappingProfile/ViewMappingProfile.js
@@ -47,6 +47,7 @@ import {
   MappingMARCBibDetails,
   MappingInvoiceDetails,
   MappingMARCAuthorityDetails,
+  MappingOrderDetails,
 } from '../detailsSections/view';
 
 import {
@@ -274,7 +275,8 @@ export class ViewMappingProfile extends Component {
       ITEM: <MappingItemDetails mappingDetails={mappingFields} />,
       INVOICE: <MappingInvoiceDetails mappingDetails={mappingFields} />,
       MARC_BIBLIOGRAPHIC: <MappingMARCBibDetails {...MARCBibDetailsProps} />,
-      MARC_AUTHORITY: <MappingMARCAuthorityDetails {...MARCBibDetailsProps} />
+      MARC_AUTHORITY: <MappingMARCAuthorityDetails {...MARCBibDetailsProps} />,
+      ORDER: <MappingOrderDetails mappingDetails={mappingFields} />,
     };
 
     return (