+++ src/settings/MappingProfiles/detailsSections/constants.js
@@ -47,3 +47,35 @@ export const WRAPPER_SOURCE_LINKS = {
   LOCATIONS: `/locations?limit=${PER_REQUEST_LIMIT}&query=${FIND_ALL_CQL} sortby name`,
   MATERIAL_TYPES: `/material-types?limit=${PER_REQUEST_LIMIT}&query=${FIND_ALL_CQL} sortby name`,
 };
+
+export const FUND_DISTRIBUTION_VISIBLE_COLUMNS = {
+  FUND_ID: 'fundId',
+  EXPENSE_CLASS_ID: 'expenseClassId',
+  VALUE: 'value',
+};
+
+export const LOCATION_VISIBLE_COLUMNS = {
+  LOCATION_ID: 'locationId',
+  QUANTITY_PHYSICAL: 'quantityPhysical',
+  QUANTITY_ELECTRONIC: 'quantityElectronic',
+};
+
+export const CONTRIBUTORS_VISIBLE_COLUMNS = {
+  CONTRIBUTOR: 'contributor',
+  CONTRIBUTOR_TYPE_ID: 'contributorNameTypeId',
+};
+
+export const PRODUCT_IDS_VISIBLE_COLUMNS = {
+  PRODUCT_ID: 'productId',
+  QUALIFIER: 'qualifier',
+  PRODUCT_ID_TYPE: 'productIdType',
+};
+
+export const NOTES_VISIBLE_COLUMNS = { NOTE: 'note' };
+
+export const PHYSICAL_RESOURCE_VISIBLE_COLUMNS = { VOLUMES: 'volumes' };
+
+export const VENDOR_VISIBLE_COLUMNS = {
+  REF_NUMBER: 'refNumber',
+  REF_NUMBER_TYPE: 'refNumberType',
+};