+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Update the Action profile UI for Orders (UIDATIMP-1231)
 * Update the Action profile UI again: Create/Edit screen (UIDATIMP-1246)
 * Data Import field mapping profile: Enable FOLIO record type = Orders (UIDATIMP-1267)
+* Create the field mapping profile for Orders and Order Lines: View (UIDATIMP-1217)
 
 ## [5.3.1](https://github.com/folio-org/ui-data-import/tree/v5.3.1) (2022-10-27)